refactor(app): extract bootstrap wiring
